refactor: make insights page more maintainable
