+++ frontend/src/insights/types/range.ts
@@ -1,0 +1,6 @@
+export type InsightsRangePreset = 'THIS_WEEK' | 'THIS_MONTH' | 'THIS_YEAR' | 'LAST_WEEK' | 'LAST_MONTH' | 'CUSTOM'
+
+export type InsightsRangeInputDates = {
+  from: string
+  to: string
+}
