+++ frontend/src/insights/utils/cashFlow.ts
@@ -1,0 +1,49 @@
+import type { InsightsCashFlowResponse } from '@/api/insights'
+import type { CashFlowBarBucket, CashFlowGranularity } from '../components/CashFlowCard'
+import {
+  getIsoWeek,
+  getMonthLabel,
+  getShortDateLabel,
+  parseYmd,
+} from './date'
+import { getCustomRangeDays } from './range'
+
+function getCashFlowGranularity(dayCount: number): CashFlowGranularity {
+  if (dayCount <= 31) return 'day'
+  if (dayCount <= 90) return 'week'
+  return 'month'
+}
+
+export function getCashFlowBarData(
+  response: InsightsCashFlowResponse | undefined,
+  fromDate: string,
+  toDate: string,
+) {
+  const dayCount = getCustomRangeDays(fromDate, toDate) ?? 1
+  const granularity = getCashFlowGranularity(dayCount)
+  return {
+    granularity,
+    buckets: (response?.points ?? []).map(([bucketStart, bucketEnd, inflow, outflow]): CashFlowBarBucket => {
+      const firstDate = parseYmd(bucketStart)
+      const lastDate = parseYmd(bucketEnd)
+      const label = granularity === 'day'
+        ? (firstDate ? getShortDateLabel(firstDate) : bucketStart)
+        : granularity === 'week'
+          ? firstDate ? `W${getIsoWeek(firstDate)}` : bucketStart
+          : firstDate ? getMonthLabel(firstDate) : bucketStart
+      const rangeLabel = firstDate && lastDate && firstDate.getTime() === lastDate.getTime()
+        ? getShortDateLabel(firstDate)
+        : firstDate && lastDate
+          ? `${getShortDateLabel(firstDate)}-${getShortDateLabel(lastDate)}`
+          : `${bucketStart}-${bucketEnd}`
+
+      return {
+        label,
+        rangeLabel,
+        inflow,
+        outflow,
+        net: inflow - outflow,
+      }
+    }),
+  }
+}
