+++ frontend/src/insights/utils/date.ts
@@ -1,0 +1,33 @@
+export function formatYmd(date: Date) {
+  const year = date.getFullYear()
+  const month = String(date.getMonth() + 1).padStart(2, '0')
+  const day = String(date.getDate()).padStart(2, '0')
+  return `${year}-${month}-${day}`
+}
+
+export function parseYmd(ymd: string) {
+  const [year, month, day] = ymd.split('-').map(Number)
+  const parsed = new Date(year, month - 1, day)
+  return Number.isNaN(parsed.getTime()) ? null : parsed
+}
+
+export function addDays(date: Date, days: number) {
+  const next = new Date(date)
+  next.setDate(next.getDate() + days)
+  return next
+}
+
+export function getShortDateLabel(date: Date) {
+  return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric' })
+}
+
+export function getMonthLabel(date: Date) {
+  return date.toLocaleDateString('en-US', { month: 'short' })
+}
+
+export function getIsoWeek(date: Date) {
+  const normalized = new Date(Date.UTC(date.getFullYear(), date.getMonth(), date.getDate()))
+  normalized.setUTCDate(normalized.getUTCDate() + 4 - (normalized.getUTCDay() || 7))
+  const yearStart = new Date(Date.UTC(normalized.getUTCFullYear(), 0, 1))
+  return Math.ceil((((normalized.getTime() - yearStart.getTime()) / 86400000) + 1) / 7)
+}
